issue-660: Fixes due to merge.
