+++ android/app/src/main/res/layout/horizontal_widget_layout.xml
