+++ android/app/src/main/res/layout/medium_widget_layout.xml
