+++ android/app/src/main/res/layout/xs_widget_layout.xml
