+++ android/app/src/main/res/xml/small_widget_provider_info.xml
@@ -1,10 +1,10 @@
 <?xml version="1.0" encoding="utf-8"?>
 <appwidget-provider xmlns:android="http://schemas.android.com/apk/res/android"
-    android:configure="fi.fmi.mobileweather.NewSmallWidgetConfigurationActivity"
-    android:initialLayout="@layout/new_small_widget_layout"
+    android:configure="fi.fmi.mobileweather.SmallWidgetConfigurationActivity"
+    android:initialLayout="@layout/small_widget_layout"
     android:minWidth="72dp"
     android:minHeight="72dp"
-    android:previewImage="@drawable/new_small_widget_preview"
+    android:previewImage="@drawable/small_widget_preview"
     android:resizeMode="horizontal|vertical"
     android:updatePeriodMillis="0"
     android:widgetCategory="home_screen">
